# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI Threat Modeler API
  description: |
    API for the AI Threat Modeler Dashboard - A comprehensive application security platform 
    with threat modeling and AI-powered chat capabilities.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token by logging in 
    via the `/api/auth/login` endpoint.
    
    ## User Roles
    - **Admin**: Full access to all features including user management and settings
    - **Operator**: Can schedule and manage their own jobs (threat modeling)
    - **Auditor**: Read-only access to view all jobs and reports
    
  version: 0.9.0
  contact:
    name: Sam Li
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: http://localhost:3001/api
    description: Local development API
...
security:
  - BearerAuth: []
paths:
...
  /api/settings:
    get:
      tags:
        - Settings
      summary: Get settings
      description: Retrieve system settings (Admin only)
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  settings:
                    $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Settings
      summary: Update settings
      description: Update system settings (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                encryption_key:
                  type: string
                  minLength: 32
                  description: Encryption key for API key storage (min 32 chars)
                anthropic_api_key:
                  type: string
                  description: Anthropic API key
                anthropic_base_url:
                  type: string
                  description: Anthropic API base URL
                  default: https://api.anthropic.com
                claude_code_max_output_tokens:
                  type: integer
                  minimum: 1
                  maximum: 1000000
                  nullable: true
                  description: Maximum output tokens for Claude Code
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Settings updated successfully
                  settings:
                    $ref: '#/components/schemas/Settings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/settings/regenerate-encryption-key:
    post:
      tags:
        - Settings
      summary: Regenerate encryption key
      description: Automatically generate a new encryption key and re-encrypt the API key (Admin only)
      responses:
        '200':
          description: Encryption key regenerated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Encryption key regenerated successfully
                  encryption_key:
                    type: string
                    description: The new encryption key (save this securely!)
                  warning:
                    type: string
                    example: Please save this encryption key securely. If you lose it, you will not be able to decrypt the API key.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/settings/validate-api-key:
    post:
      tags:
        - Settings
      summary: Validate API key
      description: Test if an Anthropic API key is valid (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - api_key
              properties:
                api_key:
                  type: string
                  description: Anthropic API key to validate
                base_url:
                  type: string
                  description: Anthropic API base URL
                  default: https://api.anthropic.com
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  message:
                    type: string
                    example: API key is valid and working correctly
                  error:
                    type: string
                    description: Error message if validation failed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
...
    Settings:
      type: object
      properties:
        encryption_key:
          type: string
          description: Encryption key for API key storage
        anthropic_api_key:
          type: string
          nullable: true
          description: Shows "***ENCRYPTED***" when API key is set
          example: "***ENCRYPTED***"
        anthropic_base_url:
          type: string
          example: https://api.anthropic.com
        claude_code_max_output_tokens:
          type: integer
          nullable: true
          minimum: 1
          maximum: 1000000
          example: 32000
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error description
        code:
          type: string
          description: Optional error code
          example: USER_NOT_FOUND
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad request
            message: Invalid parameters provided
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access token required
            message: Please provide a valid authentication token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access denied
            message: Admin role required
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            message: An unexpected error occurred